+++ src/models/ers_core.py
@@ -91,167 +91,232 @@ linkml_meta = LinkMLMeta({'default_prefix': 'ers',
      'source_file': 'resources/schema/ers-core_v0.1.0.yaml'} )
 
 
-class RequestOrResponseMixin(ConfiguredBaseModel):
+class ERECommunicationArtefact(ConfiguredBaseModel):
     """
-    Root mixin to represent attributes common to both requests and results.
+    Root abstraction to represent attributes common to both requests and results.
+    This is modelled as a mixin in LinkML (so that it can't be instantiated directly).
 
     """
     linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'abstract': True,
          'from_schema': 'https://data.europa.eu/ers/schema',
          'mixin': True})
 
-    type: Literal["RequestOrResponseMixin"] = Field(default="RequestOrResponseMixin", description="""The type of the request or result.
+    type: Literal["ERECommunicationArtefact"] = Field(default="ERECommunicationArtefact", description="""The type of the request or result.
 
 As per LinkML specification, `designates_type` is used here in order to allow for this
 slot to tell the concrete subclass that an instance (such as a JSON object) belongs to.
 
 In other words, a particular request will have `type` set with values like 
-`EntityResolutionRequest` or `EntityResolutionResult`
-""", json_schema_extra = { "linkml_meta": {'designates_type': True, 'domain_of': ['RequestOrResponseMixin', 'Entity']} })
+`EntityMentionResolutionRequest` or `EntityResolutionResult`
+""", json_schema_extra = { "linkml_meta": {'designates_type': True,
+         'domain_of': ['ERECommunicationArtefact', 'EntityMention']} })
     metadata: Optional[str] = Field(default=None, description="""An optional arbitrary dictionary of further request metadata.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['RequestOrResponseMixin']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERECommunicationArtefact']} })
 
 
-class Request(RequestOrResponseMixin):
+class ERERequest(ERECommunicationArtefact):
     """
     Root class to represent all the requests sent to the ERE.
 
     """
     linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'abstract': True,
          'from_schema': 'https://data.europa.eu/ers/schema',
-         'mixins': ['RequestOrResponseMixin']})
+         'mixins': ['ERECommunicationArtefact']})
 
     requestId: str = Field(default=..., description="""A string representing the unique ID of this request.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Request', 'Response']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest', 'EREResponse']} })
     originator: str = Field(default=..., description="""The ID or URI of the request originator.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Request']} })
-    type: Literal["Request"] = Field(default="Request", description="""The type of the request or result.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest']} })
+    creationTime: Optional[datetime ] = Field(default=None, description="""The timestamp when the request was created.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest']} })
+    type: Literal["ERERequest"] = Field(default="ERERequest", description="""The type of the request or result.
 
 As per LinkML specification, `designates_type` is used here in order to allow for this
 slot to tell the concrete subclass that an instance (such as a JSON object) belongs to.
 
 In other words, a particular request will have `type` set with values like 
-`EntityResolutionRequest` or `EntityResolutionResult`
-""", json_schema_extra = { "linkml_meta": {'designates_type': True, 'domain_of': ['RequestOrResponseMixin', 'Entity']} })
+`EntityMentionResolutionRequest` or `EntityResolutionResult`
+""", json_schema_extra = { "linkml_meta": {'designates_type': True,
+         'domain_of': ['ERECommunicationArtefact', 'EntityMention']} })
     metadata: Optional[str] = Field(default=None, description="""An optional arbitrary dictionary of further request metadata.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['RequestOrResponseMixin']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERECommunicationArtefact']} })
 
 
-class Response(RequestOrResponseMixin):
+class EREResponse(ERECommunicationArtefact):
     """
     Root class to represent all the responses sent by the ERE.
 
     """
     linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'abstract': True,
          'from_schema': 'https://data.europa.eu/ers/schema',
-         'mixins': ['RequestOrResponseMixin']})
+         'mixins': ['ERECommunicationArtefact']})
 
     requestId: str = Field(default=..., description="""A string representing the unique ID of the request this response is about.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Request', 'Response']} })
-    type: Literal["Response"] = Field(default="Response", description="""The type of the request or result.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest', 'EREResponse']} })
+    type: Literal["EREResponse"] = Field(default="EREResponse", description="""The type of the request or result.
 
 As per LinkML specification, `designates_type` is used here in order to allow for this
 slot to tell the concrete subclass that an instance (such as a JSON object) belongs to.
 
 In other words, a particular request will have `type` set with values like 
-`EntityResolutionRequest` or `EntityResolutionResult`
-""", json_schema_extra = { "linkml_meta": {'designates_type': True, 'domain_of': ['RequestOrResponseMixin', 'Entity']} })
+`EntityMentionResolutionRequest` or `EntityResolutionResult`
+""", json_schema_extra = { "linkml_meta": {'designates_type': True,
+         'domain_of': ['ERECommunicationArtefact', 'EntityMention']} })
     metadata: Optional[str] = Field(default=None, description="""An optional arbitrary dictionary of further request metadata.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['RequestOrResponseMixin']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERECommunicationArtefact']} })
 
 
-class EntityResolutionRequest(Request):
+class EntityMentionResolutionRequest(ERERequest):
     """
     An entity resolution request sent to the ERE, containing the entity to be resolved.
 
     """
-    linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'examples': [{'value': '{\n'
-                                '  "type": "EntityResolutionRequest",            \n'
-                                '  "entity": \n'
+    linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'examples': [{'description': 'a regular request',
+                       'value': '{\n'
+                                '  "type": '
+                                '"EntityMentionResolutionRequest",            \n'
+                                '  "entityMention": \n'
                                 '  { \n'
                                 '    "type": "http://www.w3.org/ns/org#Organization",\n'
-                                '    "id": '
+                                '    // As said above, there is always a way to '
+                                'compute this\n'
+                                '    "identifier": '
                                 '"http://data.europa.eu/ers/id/324fs3r345vx-aa32wa",\n'
-                                '    "entityData": "epd:ent005 a org:Organization; '
-                                '...   cccev:telephone \\"+44 1924306780\\" .",\n'
-                                '    "entityDataFormat": "text/turtle"\n'
+                                '    "payload": "epd:ent005 a org:Organization; ...   '
+                                'cccev:telephone \\"+44 1924306780\\" .",\n'
+                                '    "dataFormat": "text/turtle",\n'
+                                '    // As Said above, this is optional and JSON-LD is '
+                                'just an example of what it could rendered.\n'
+                                '    "jsonRepresentation": {\n'
+                                '      "@context": {\n'
+                                '        "org": "http://www.w3.org/ns/org#",\n'
+                                '        "cccev": '
+                                '"https://data.europa.eu/cc/cefact/code/"\n'
+                                '      },\n'
+                                '      "@id": '
+                                '"http://data.europa.eu/ers/id/324fs3r345vx-aa32wa",\n'
+                                '      "@type": "org:Organization",\n'
+                                '      "cccev:telephone": "+44 1924306780"\n'
+                                '    }\n'
                                 '  },\n'
                                 '  "requestId": "324fs3r345vx",\n'
                                 '  "originator": "TED SWS pipeline",\n'
+                                '  "creationTime": "2026-01-14T12:34:56Z",\n'
                                 '  "metadata": {\n'
                                 '    "originator system": "VocBench editor",\n'
                                 '    "originator timestamp": "23748737643"\n'
                                 '  }\n'
+                                '}\n'},
+                      {'description': 'a refresh request (ie, carrying a rejection '
+                                      'list)',
+                       'value': '{\n'
+                                '  "type": '
+                                '"EntityMentionResolutionRequest",            \n'
+                                '  "entityMention": \n'
+                                '  { \n'
+                                '    "type": "http://www.w3.org/ns/org#Organization",\n'
+                                '    "identifier": '
+                                '"http://data.europa.eu/ers/id/324fs3r345vx-aa32wa",\n'
+                                '    "payload": "epd:ent005 a org:Organization; ...   '
+                                'cccev:telephone \\"+44 1924306780\\" .",\n'
+                                '    "dataFormat": "text/turtle"\n'
+                                '  },\n'
+                                '  "rejectedCanonicalIdentifiers": [\n'
+                                '    '
+                                '"http://data.europa.eu/ers/id/324fs3r345vx-bb45we",\n'
+                                '    '
+                                '"http://data.europa.eu/ers/id/324fs3r345vx-cc67ui"\n'
+                                '  ],\n'
+                                '  "requestId": "324fs3r345vx01",\n'
+                                '  "originator": "TED SWS pipeline",\n'
+                                '  "creationTime": "2026-01-14T12:40:56Z"\n'
                                 '}\n'}],
          'from_schema': 'https://data.europa.eu/ers/schema'})
 
-    entity: Entity = Field(default=..., description="""The data about the entity to be resolved.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['EntityResolutionRequest']} })
+    entityMention: EntityMention = Field(default=..., description="""The data about the entity to be resolved.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['EntityMentionResolutionRequest']} })
+    rejectedCanonicalIdentifiers: Optional[list[str]] = Field(default=[], description="""When this is present, the request is a refresh request: it is asking that the entity 
+is resolved again and the clusters/canonical entities that were previously proposed 
+as resolution are now ignored.
+
+The exact reaction to this is implementation dependent. In the simplest case, the ERE
+might just create a singleton cluster with this entity as member. In a more advanced 
+case, it might recompute the similarity with more advanced algorithms or use updated
+data.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['EntityMentionResolutionRequest']} })
     requestId: str = Field(default=..., description="""A string representing the unique ID of this request.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Request', 'Response']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest', 'EREResponse']} })
     originator: str = Field(default=..., description="""The ID or URI of the request originator.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Request']} })
-    type: Literal["EntityResolutionRequest"] = Field(default="EntityResolutionRequest", description="""The type of the request or result.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest']} })
+    creationTime: Optional[datetime ] = Field(default=None, description="""The timestamp when the request was created.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest']} })
+    type: Literal["EntityMentionResolutionRequest"] = Field(default="EntityMentionResolutionRequest", description="""The type of the request or result.
 
 As per LinkML specification, `designates_type` is used here in order to allow for this
 slot to tell the concrete subclass that an instance (such as a JSON object) belongs to.
 
 In other words, a particular request will have `type` set with values like 
-`EntityResolutionRequest` or `EntityResolutionResult`
-""", json_schema_extra = { "linkml_meta": {'designates_type': True, 'domain_of': ['RequestOrResponseMixin', 'Entity']} })
+`EntityMentionResolutionRequest` or `EntityResolutionResult`
+""", json_schema_extra = { "linkml_meta": {'designates_type': True,
+         'domain_of': ['ERECommunicationArtefact', 'EntityMention']} })
     metadata: Optional[str] = Field(default=None, description="""An optional arbitrary dictionary of further request metadata.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['RequestOrResponseMixin']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERECommunicationArtefact']} })
 
 
-class EntityResolutionResponse(Response):
+class EntityMentionResolutionResponse(EREResponse):
     """
     An entity resolution response sent by the ERE.
 
-    This contains a reference to the canonical entity that the ERE has associated to the original 
-    entity in the request. It also reports a confidence score for the established association.
+    This links an `AlignmentLinkSet`, to represent possible resolutions (see the attribute definition).
 
     """
     linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'examples': [{'value': '{\n'
-                                '  "type": "EntityResolutionResponse",\n'
-                                '  "sourceEntityId": '
+                                '  "type": "EntityMentionResolutionResponse",\n'
+                                '  "requestId": "324fs3r345vx",\n'
+                                '  "alignmentLinkSet": {\n'
+                                '    "subjectEntityMentionIdentifier": '
                                 '"http://data.europa.eu/ers/id/324fs3r345vx-q11rea",\n'
-                                '  "confidenceLevel": 0.91,\n'
-                                '  "requestId": "324fs3r345vx"\n'
-                                '  "canonicalEntity": \n'
-                                '  { \n'
-                                '    "type": "http://www.w3.org/ns/org#Organization",\n'
-                                '    "id": '
+                                '    "alignmentOptions": [\n'
+                                '      { \n'
+                                '        "canonicalIdentifier": '
                                 '"http://data.europa.eu/ers/id/324fs3r345vx-aa32wa",\n'
-                                '    "entityData": "epd:ent001 a org:Organization; '
-                                '...   cccev:telephone \\"+441924306780\\" .",\n'
-                                '    "entityDataFormat": "text/turtle"\n'
+                                '        "confidenceScore": 0.91\n'
+                                '      },\n'
+                                '      { \n'
+                                '        "canonicalIdentifier": '
+                                '"http://data.europa.eu/ers/id/324fs3r345vx-bb45we",\n'
+                                '        "confidenceScore": 0.65\n'
+                                '      }\n'
+                                '    ]\n'
                                 '  }\n'
-                                '}\n'}],
+                                '}\n'
+                                '    \n'}],
          'from_schema': 'https://data.europa.eu/ers/schema'})
 
-    canonicalEntity: CanonicalEntity = Field(default=..., description="""The canonical entity that the ERE has associated to the original entity.
-This includes the canonical entity URI and its type.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['EntityResolutionResponse']} })
-    sourceEntityId: str = Field(default=..., description="""The ID or URI of the original entity that has been resolved.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['EntityResolutionResponse']} })
-    confidenceLevel: Optional[float] = Field(default=None, description="""A 0-1 value of how confident the ERE is about associating the original entity
-with the canonical entity's cluster.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['EntityResolutionResponse']} })
+    alignmentLinkSet: AlignmentLinkSet = Field(default=..., description="""The set of alignment links representing the candidate canonical entities/clusters
+that the entity mention in the original request could align to (be equivalent to).
+
+**Note**: for the moment, this is not multi-valued, since we don't support batch requests (yet?),
+thus there is only one set in a response, that resolves for the single entity mention in the
+original request (with multiple alignment candidates). If, in the future, we support batch requests,
+then we might need to return one alignment link set per entity mention in a request.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['EntityMentionResolutionResponse']} })
     requestId: str = Field(default=..., description="""A string representing the unique ID of the request this response is about.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Request', 'Response']} })
-    type: Literal["EntityResolutionResponse"] = Field(default="EntityResolutionResponse", description="""The type of the request or result.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest', 'EREResponse']} })
+    type: Literal["EntityMentionResolutionResponse"] = Field(default="EntityMentionResolutionResponse", description="""The type of the request or result.
 
 As per LinkML specification, `designates_type` is used here in order to allow for this
 slot to tell the concrete subclass that an instance (such as a JSON object) belongs to.
 
 In other words, a particular request will have `type` set with values like 
-`EntityResolutionRequest` or `EntityResolutionResult`
-""", json_schema_extra = { "linkml_meta": {'designates_type': True, 'domain_of': ['RequestOrResponseMixin', 'Entity']} })
+`EntityMentionResolutionRequest` or `EntityResolutionResult`
+""", json_schema_extra = { "linkml_meta": {'designates_type': True,
+         'domain_of': ['ERECommunicationArtefact', 'EntityMention']} })
     metadata: Optional[str] = Field(default=None, description="""An optional arbitrary dictionary of further request metadata.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['RequestOrResponseMixin']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERECommunicationArtefact']} })
 
 
-class ErrorResponse(Response):
+class EREErrorResponse(EREResponse):
     """
     Response sent by the ERE when some error/exception occurs while processing a request.
     For instance, this may happen if the request is malformed or some internal error happens.
@@ -260,14 +325,14 @@ class ErrorResponse(Response):
 
     """
     linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'examples': [{'value': '{\n'
-                                '  "type": "ErrorResponse",\n'
+                                '  "type": "EREErrorResponse",\n'
                                 '  "requestId": "324fs3r345vx",\n'
                                 '  "errorType": '
                                 '"ere.exceptions.MalformedRequestError",\n'
                                 '  "errorTitle": "The entity data is missing in the '
                                 'request",\n'
                                 '  "errorDetail": "The \'entity\' attribute is '
-                                'required in EntityResolutionRequest message",\n'
+                                'required in EntityMentionResolutionRequest message",\n'
                                 '  // Optional and not recommended for production use\n'
                                 '  "errorTrace": "Traceback (most recent call '
                                 'last):\\n  File \\"/app/ere/service.py\\", line 45, '
@@ -278,91 +343,119 @@ class ErrorResponse(Response):
     errorType: str = Field(default=..., description="""A string representing the error type, eg, the FQN of the raised exception.
 
 This corresponds to RFC-9457's `type`.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['ErrorResponse']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['EREErrorResponse']} })
     errorTitle: Optional[str] = Field(default=None, description="""A human readable brief message about the error that occurred.
 
 This corresponds to RFC-9457's `title`.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['ErrorResponse']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['EREErrorResponse']} })
     errorDetail: Optional[str] = Field(default=None, description="""A human readable detailed message about the error that occurred.
 
 This corresponds to RFC-9457's `detail`.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['ErrorResponse']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['EREErrorResponse']} })
     errorTrace: Optional[str] = Field(default=None, description="""A string representing a (stack) trace of the error that occurred.
 
 This is optional and typically used for debugging purposes only, since
 exposing this kind of server-side information is a security risk.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['ErrorResponse']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['EREErrorResponse']} })
     requestId: str = Field(default=..., description="""A string representing the unique ID of the request this response is about.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Request', 'Response']} })
-    type: Literal["ErrorResponse"] = Field(default="ErrorResponse", description="""The type of the request or result.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest', 'EREResponse']} })
+    type: Literal["EREErrorResponse"] = Field(default="EREErrorResponse", description="""The type of the request or result.
 
 As per LinkML specification, `designates_type` is used here in order to allow for this
 slot to tell the concrete subclass that an instance (such as a JSON object) belongs to.
 
 In other words, a particular request will have `type` set with values like 
-`EntityResolutionRequest` or `EntityResolutionResult`
-""", json_schema_extra = { "linkml_meta": {'designates_type': True, 'domain_of': ['RequestOrResponseMixin', 'Entity']} })
+`EntityMentionResolutionRequest` or `EntityResolutionResult`
+""", json_schema_extra = { "linkml_meta": {'designates_type': True,
+         'domain_of': ['ERECommunicationArtefact', 'EntityMention']} })
     metadata: Optional[str] = Field(default=None, description="""An optional arbitrary dictionary of further request metadata.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['RequestOrResponseMixin']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERECommunicationArtefact']} })
 
 
-class Entity(ConfiguredBaseModel):
+class EntityMention(ConfiguredBaseModel):
     """
-    An entity is a representation of a real-world entity, as provided by the ERS.
-    It contains the entity data (e.g. RDF description) along with metadata about
-    the entity, such as its type and the data format used to represent it.
+    An entity mention is a representation of a real-world entity in the ERS. It must have 
+    a data content and a data format, so that components like the ERE can use them for resolution.
 
-    """
-    linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'from_schema': 'https://data.europa.eu/ers/schema'})
-
-    id: Optional[str] = Field(default=None, description="""A string containing the entity ID or URI (set by the ERS or, for canonical entities, by the ERE).
-
-Note that the ID isn't mandatory when an entity is submitted for resolution, since the initial input
-might be something like unstructured text, where the entity and its ID is to be recognised.        
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Entity', 'CanonicalEntity']} })
-    type: str = Field(default=..., description="""A string representing the entity type URI (based on CET).
-
-Note that we don't use the `designates_type` thing here, since entities or canonical entities 
-are always used in clearly distinct contexts.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['RequestOrResponseMixin', 'Entity']} })
-    entityDataFormat: Optional[str] = Field(default=None, description="""A string about the MIME format of `entityData` (e.g. text/turtle, application/ld+json)
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Entity']} })
-    entityData: Optional[str] = Field(default=None, description="""A code string representing the entity details (eg, RDF description).
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Entity']} })
-
-
-class CanonicalEntity(Entity):
-    """
-    A canonical entity is an entity that the ERE has created during the resolution process
-    of ERS entities.
-
-    TODO: we don't support lineage for the moment, see the ERE contract document.
+    Moreover, an entity mention must have a computed identifier (see below).
 
     """
     linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'from_schema': 'https://data.europa.eu/ers/schema'})
 
-    id: str = Field(default=..., description="""The (canonical) URI of the canonical entity. This restricts the parent range to URIs only.
+    identifier: str = Field(default=..., description="""An URI identifying the entity.
 
-Contrary to `Entity.id`, this is always known/required for canonical entities.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Entity', 'CanonicalEntity']} })
+While mandatory, this can be computed, using same function that depends on the entity payload.
+In that case, **there must be** a single function in the whole ERS (including the ERE) that
+computes the same identifier for the same payload, eg, a hash, an RDF URI extractor. This is
+needed for the resolution results to refer to the correct request entities.        
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['EntityMention']} })
     type: str = Field(default=..., description="""A string representing the entity type URI (based on CET).
 
 Note that we don't use the `designates_type` thing here, since entities or canonical entities 
 are always used in clearly distinct contexts.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['RequestOrResponseMixin', 'Entity']} })
-    entityDataFormat: Optional[str] = Field(default=None, description="""A string about the MIME format of `entityData` (e.g. text/turtle, application/ld+json)
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Entity']} })
-    entityData: Optional[str] = Field(default=None, description="""A code string representing the entity details (eg, RDF description).
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Entity']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERECommunicationArtefact', 'EntityMention']} })
+    datFormat: Optional[str] = Field(default=None, description="""A string about the MIME format of `payload` (e.g. text/turtle, application/ld+json)
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['EntityMention']} })
+    payload: Optional[str] = Field(default=None, description="""A code string representing the entity details (eg, RDF description).
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['EntityMention']} })
+    jsonRepresentation: Optional[str] = Field(default=None, description="""An optional JSON representation of the entity, which is usually achieved from the payload.
+This is mainly useful for the curation app.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['EntityMention']} })
 
 
-class RebuildRequest(Request):
+class AlignmentLinkSet(ConfiguredBaseModel):
+    """
+    A set of alignment links to a referred entity.
+
+    Each link in the set represents an entity that might be equivalent to the referred entity,
+    see `AlignmentLink` for details.
+
+    """
+    linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'from_schema': 'https://data.europa.eu/ers/schema'})
+
+    subjectEntityMentionIdentifier: str = Field(default=..., description="""The identifier of the entity mention that is the subject of these alignment links.
+This must match the `identifier` attribute of an `EntityMention` that a resolution response
+refers to.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['AlignmentLinkSet']} })
+    alignmentOptions: list[AlignmentLink] = Field(default=..., description="""A list of possible matches (alignment links) between the subject entity mention
+and candidate canonical entities.
+
+It is recommended that these are sorted by descending confidence score, although
+that is not mandatory.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['AlignmentLinkSet']} })
+
+
+class AlignmentLink(ConfiguredBaseModel):
+    """
+    An alignment link representing a possible equivalence between an entity mention in the
+    `AlignmentLinkSet` the link belongs to, and a canonical entity, together with a confidence score.
+
+    A semi-formal representation:
+
+    ```
+      for each (canonicalIdentifier, cconfidenceScore) in AlignmentLinkSet.alignmentOptions:
+        entity(subjectEntityMentionIdentifier)  ==  entity(mentionIdentifier) 
+          with score = confidenceScore
+    ```
+
+    """
+    linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'from_schema': 'https://data.europa.eu/ers/schema'})
+
+    canonicalIdentifier: str = Field(default=..., description="""The identifier of the cluster/canonical entity that is considered equivalent to the
+subject entity mention in the `AlignmentLinkSet` the link belongs to.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['AlignmentLink']} })
+    confidenceScore: float = Field(default=..., description="""A 0-1 value of how confident the ERE is about the equivalence between the subject entity mention
+and the target canonical entity.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['AlignmentLink']} })
+
+
+class FullRebuildRequest(ERERequest):
     """
     A request to reset all the resolutions computed so far and rebuild them as 
     requests about old entities arrive again (and build new entities from scratch).
 
     It is expected that the ERE client re-sends all the entities to be resolved again,
-    using `EntityResolutionRequest` messages exactly as the first time the resolutions 
+    using `EntityMentionResolutionRequest` messages exactly as the first time the resolutions 
     were built. This implies the a client like the ERS logs/persists the entities it receives
     to resolve and also saves manual overriding of ERE results.
 
@@ -370,24 +463,27 @@ class RebuildRequest(Request):
     linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'from_schema': 'https://data.europa.eu/ers/schema'})
 
     requestId: str = Field(default=..., description="""A string representing the unique ID of this request.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Request', 'Response']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest', 'EREResponse']} })
     originator: str = Field(default=..., description="""The ID or URI of the request originator.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Request']} })
-    type: Literal["RebuildRequest"] = Field(default="RebuildRequest", description="""The type of the request or result.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest']} })
+    creationTime: Optional[datetime ] = Field(default=None, description="""The timestamp when the request was created.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest']} })
+    type: Literal["FullRebuildRequest"] = Field(default="FullRebuildRequest", description="""The type of the request or result.
 
 As per LinkML specification, `designates_type` is used here in order to allow for this
 slot to tell the concrete subclass that an instance (such as a JSON object) belongs to.
 
 In other words, a particular request will have `type` set with values like 
-`EntityResolutionRequest` or `EntityResolutionResult`
-""", json_schema_extra = { "linkml_meta": {'designates_type': True, 'domain_of': ['RequestOrResponseMixin', 'Entity']} })
+`EntityMentionResolutionRequest` or `EntityResolutionResult`
+""", json_schema_extra = { "linkml_meta": {'designates_type': True,
+         'domain_of': ['ERECommunicationArtefact', 'EntityMention']} })
     metadata: Optional[str] = Field(default=None, description="""An optional arbitrary dictionary of further request metadata.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['RequestOrResponseMixin']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERECommunicationArtefact']} })
 
 
-class RebuildResponse(Response):
+class FullRebuildResponse(EREResponse):
     """
-    A response to a `RebuildRequest`, confirming that the rebuild process has started.
+    A response to a `FullRebuildRequest`, confirming that the rebuild process has started.
 
     This should carry the `requestId` attribute.
 
@@ -395,28 +491,30 @@ class RebuildResponse(Response):
     linkml_meta: ClassVar[LinkMLMeta] = LinkMLMeta({'from_schema': 'https://data.europa.eu/ers/schema'})
 
     requestId: str = Field(default=..., description="""A string representing the unique ID of the request this response is about.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['Request', 'Response']} })
-    type: Literal["RebuildResponse"] = Field(default="RebuildResponse", description="""The type of the request or result.
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERERequest', 'EREResponse']} })
+    type: Literal["FullRebuildResponse"] = Field(default="FullRebuildResponse", description="""The type of the request or result.
 
 As per LinkML specification, `designates_type` is used here in order to allow for this
 slot to tell the concrete subclass that an instance (such as a JSON object) belongs to.
 
 In other words, a particular request will have `type` set with values like 
-`EntityResolutionRequest` or `EntityResolutionResult`
-""", json_schema_extra = { "linkml_meta": {'designates_type': True, 'domain_of': ['RequestOrResponseMixin', 'Entity']} })
+`EntityMentionResolutionRequest` or `EntityResolutionResult`
+""", json_schema_extra = { "linkml_meta": {'designates_type': True,
+         'domain_of': ['ERECommunicationArtefact', 'EntityMention']} })
     metadata: Optional[str] = Field(default=None, description="""An optional arbitrary dictionary of further request metadata.
-""", json_schema_extra = { "linkml_meta": {'domain_of': ['RequestOrResponseMixin']} })
+""", json_schema_extra = { "linkml_meta": {'domain_of': ['ERECommunicationArtefact']} })
 
 
 # Model rebuild
 # see https://pydantic-docs.helpmanual.io/usage/models/#rebuilding-a-model
-RequestOrResponseMixin.model_rebuild()
-Request.model_rebuild()
-Response.model_rebuild()
-EntityResolutionRequest.model_rebuild()
-EntityResolutionResponse.model_rebuild()
-ErrorResponse.model_rebuild()
-Entity.model_rebuild()
-CanonicalEntity.model_rebuild()
-RebuildRequest.model_rebuild()
-RebuildResponse.model_rebuild()
+ERECommunicationArtefact.model_rebuild()
+ERERequest.model_rebuild()
+EREResponse.model_rebuild()
+EntityMentionResolutionRequest.model_rebuild()
+EntityMentionResolutionResponse.model_rebuild()
+EREErrorResponse.model_rebuild()
+EntityMention.model_rebuild()
+AlignmentLinkSet.model_rebuild()
+AlignmentLink.model_rebuild()
+FullRebuildRequest.model_rebuild()
+FullRebuildResponse.model_rebuild()