+++ CHANGELOG.md
@@ -1,6 +1,25 @@
-# Change log
+# Changelog
 
-## 0.1.0 - 2025-12-22
+All notable changes to this project will be documented in this file.
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
+
+## [Unreleased]
+### Added
+
+### Changed
+
+
+## [0.1.0-rc.2] - 2026-01-16
+### Added
+* data model: Support for rejected canonical identifiers
+* data model: Support for a full rebuild
+
+### Changed
+* Updated and refined data model
+* Updated ERS-ERE  examples
+
+
+## [0.1.0-rc.1] - 2025-12-22
 
 * Initial release, fulfilling Project's Delivery 1 (ERE Technical Contract and related code).
 * [LinkML schema](resources/schema/ers-core_v0.1.0.yaml) to specify the interaction with the ERE service.